separate filename and metadata by tabs 🤦🏻‍♀️
soooooooooooooo thank you @keith for raising this in #11. turns out _every_ sample diff uses a tab character to separate the filename. git definitely always does. i haven't yet found any counter-examples (except those generated from this library 🤦🏻‍♀️🤦🏻‍♀️🤦🏻‍♀️).

well, there's one very annoying counter-example: one of the examples on gnu.org's written description of the format.

- full example (uses tabs): https://www.gnu.org/software/diffutils/manual/html_node/Example-Unified.html
- detailed description (uses spaces, doesn't specify): https://www.gnu.org/software/diffutils/manual/html_node/Detailed-Unified.html

i'm going to assume that the detailed description contains an error and ignore it 🙃. this library will switch to expect (and render) a tab character. not sure how else to handle it, but if new counter-examples come up, i'll think about thinking more about it with thoughts.

-----

would be so nice if there was an actual spec for diffs. and other fantasies.